Initial GL work. Trying to port X11 to GLU (pcsxr existing plugin)
Note that dithering is now a bool (default off)
Seems to be a linking error or something... Not sure, but it keeps saying that functions are undefined. I need to dig into the code further.
GLU is heavily unimplemented, just some random code tossed in right now.
I'm also dropping support for Gamecube, as this will be eventually ported to libwiigui